let the menu determine where the page begins
Remove this header nonsense, where the menu is floating and then
an additional header element is added for spacing. This causes
the menu to flow into the text as soon as the menu becomes
 multi-lined.

Insted the menu should not be positioned absolute, so it can
determine it's own space and the body container is pushed down
accordingly.