prevent the menu change to move around content
when you scoll down then the menu becomes floating. With my latest
layout changes, this left a gap and the content would move up. When
you scroll up again the menu overlays the body.

This was fixed by not making the menu absolut after scrolling up.

However this fix still leaves us with an unpleasant bouncing artifact,
where the container moves up/down when the menu changes from being
at the top of the page, to the top of the screen. This commit fixes
the movement by teporarily filling the gap left by the menu with an
empty div of the same size.